requirements.txt
Flask==3.1.1
Flask-CORS==5.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
gunicorn==23.0.0

